+++ src/main/java/frc/robot/commands/LEDs/BlinkSignalLight.java
@@ -2,25 +2,29 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.LEDs;
 
 import frc.robot.subsystems.LEDSubsystem;
+import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 
 /** An example command that uses an example subsystem. */
-public class SetSignalLightIntensity extends Command {
+public class BlinkSignalLight extends Command {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
   private final LEDSubsystem m_LEDSubsystem;
   double intensity;
+  Timer timey = new Timer();
+  Double Delay;
 
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public SetSignalLightIntensity(LEDSubsystem subsystem, double input) {
+  public BlinkSignalLight(LEDSubsystem subsystem, double IntensityInput, double DelayInput) {
     m_LEDSubsystem = subsystem;
-    intensity = input;
+    intensity = IntensityInput;
+    Delay = DelayInput;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(m_LEDSubsystem);
   }
@@ -31,21 +35,31 @@ public class SetSignalLightIntensity extends Command {
   @Override
   public void initialize() {
     m_LEDSubsystem.SetLightIntensity(0);
+    timey.start();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_LEDSubsystem.SetLightIntensity(intensity);
+    if (timey.hasElapsed(Delay)) {
+      if (m_LEDSubsystem.GetLightIntensity() < intensity/2) {
+        m_LEDSubsystem.SetLightIntensity(intensity);
+      } else {
+        m_LEDSubsystem.SetLightIntensity(0);
+      }
+      timey.restart();
+    }
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_LEDSubsystem.SetLightIntensity(0);
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return true;
+    return false;
   }
 }
